Moving around the job subclasses.
